+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Partner-id comparison made case insensitive
 - Reverted from NNG to nanomag (v1.1.2)
 - reverted temporary CMake reference to https://github.com/bill1600/seshat
+- Added log for time difference of parodus connect time and boot time
 
 ## [1.0.1] - 2018-07-18
 ### Added